backend/requirements.txt
Flask==2.1.0
requests==2.26.0
jsonschema==3.2.0
psycopg2-binary==2.9.1


